+++ src/maou/app/pre_process/feature.py
@@ -26,7 +26,7 @@ def make_feature(board: cshogi.Board) -> np.ndarray:  # type: ignore
     Returns:
         Feature array with shape (FEATURES_NUM, 9, 9)
     """
-    features = np.empty((FEATURES_NUM, 9, 9), dtype=np.float32)
+    features = np.empty((FEATURES_NUM, 9, 9), dtype=np.uint8)
     features.fill(0)
     if board.turn == cshogi.BLACK:  # type: ignore
         board.piece_planes(features)
